+++ src/Maybe/expect.ts
@@ -1,13 +1,6 @@
-import { isNotNullAndUndefined, Maybe, NotNullAndUndefined } from './Maybe.js';
-
 /**
- *  Return _input_ as `T` if the passed _input_ is not `null` and `undefined`.
- *  Otherwise, throw `TypeError` with the passed `msg`.
+ *  @deprecated
+ *  This module is kept for backward compatibility.
+ *  Please use `option-t/Maybe/Maybe` directly.
  */
-export function expectNotNullAndUndefined<T>(input: Maybe<T>, msg: string): NotNullAndUndefined<T> {
-    if (isNotNullAndUndefined(input)) {
-        return input;
-    }
-
-    throw new TypeError(msg);
-}
+export { expectNotNullAndUndefined } from './Maybe.js';
