+++ src/Maybe/index.ts
@@ -1,9 +1,15 @@
-export { NotNullAndUndefined, Maybe, isNotNullAndUndefined, isNullOrUndefined } from './Maybe.js';
+export {
+    type NotNullAndUndefined,
+    type Maybe,
+    isNotNullAndUndefined,
+    isNullOrUndefined,
+    expectNotNullAndUndefined,
+    unwrapMaybe,
+} from './Maybe.js';
 // XXX: `and()` operation is equivalent of `a && b` so we don't ship it by default set.
 //export { andForMaybe as and } from './and.js';
 export { andThenForMaybe as andThen } from './andThen.js';
 export { andThenAsyncForMaybe as andThenAsync } from './andThenAsync.js';
-export { expectNotNullAndUndefined as expect } from './expect.js';
 export { inspectMaybe as inspect } from './inspect.js';
 export { mapForMaybe as map } from './map.js';
 export { mapAsyncForMaybe as mapAsync } from './mapAsync.js';
@@ -15,7 +21,22 @@ export { mapOrElseAsyncForMaybe as mapOrElseAsync } from './mapOrElseAsync.js';
 // export { orForMaybe as or } from './or.js';
 export { orElseForMaybe as orElse } from './orElse.js';
 export { orElseAsyncForMaybe as orElseAsync } from './orElseAsync.js';
-export { unwrapMaybe as unwrap } from './unwrap.js';
 export { unwrapOrFromMaybe as unwrapOr } from './unwrapOr.js';
 export { unwrapOrElseFromMaybe as unwrapOrElse } from './unwrapOrElse.js';
 export { unwrapOrElseAsyncFromMaybe as unwrapOrElseAsync } from './unwrapOrElseAsync.js';
+
+import { expectNotNullAndUndefined, unwrapMaybe } from './Maybe.js';
+
+/**
+ *  @deprecated
+ *  This is an alias for backward compatibility.
+ *  Please use {@link expectNotNullAndUndefined} instead.
+ */
+export const expect: typeof expectNotNullAndUndefined = expectNotNullAndUndefined;
+
+/**
+ *  @deprecated
+ *  This is an alias for backward compatibility.
+ *  Please use {@link unwrapMaybe} instead.
+ */
+export const unwrap: typeof unwrapMaybe = unwrapMaybe;
